Rename PaymentFeeAdjustmentCleaner to PaymentFeeAdjustmentClearer
